+++ handsontable/types/editors/dateEditor/index.d.ts
@@ -1,0 +1,1 @@
+export { EDITOR_TYPE, DateEditor } from './dateEditor';