+++ handsontable/types/editors/numericEditor/index.d.ts
@@ -1,0 +1,1 @@
+export { EDITOR_TYPE, NumericEditor } from './numericEditor';